Testing
Still broken on Linux

+++ Common/veldrid/src/Veldrid.MetalBindings/CGFloat.cs
@@ -1,3 +1,5 @@
+using System;
+using System.Runtime.CompilerServices;
 
 namespace Veldrid.MetalBindings
 {
@@ -17,8 +19,7 @@ namespace Veldrid.MetalBindings
             get => _value;
         }
 
-        public static implicit operator CGFloat(double value) => new(value);
-
+        public static implicit operator CGFloat(double value) => new CGFloat(value);
         public static implicit operator double(CGFloat cgf) => cgf.Value;
 
         public override string ToString() => _value.ToString();
